introduce unified base class for local llm models
